requirements.txt
aiosqlite~=0.20.0
python-dateutil~=2.9.0
python-dotenv~=1.0.1
twitchio~=2.9.1
